+++ kiltergpt/data/datasets.py
@@ -1,28 +1,32 @@
+import math
 from pathlib import Path
 
 import pandas as pd
 import torch
+import torch.nn.functional as F
 from torch.utils.data import Dataset
 
 from .tokenizer import Tokenizer
 
 
-class KilterGPTDataset(Dataset):
+class KilterDataset(Dataset):
     def __init__(
         self,
         filename: str | Path,
         tokenizer: Tokenizer,
         *,
-        context_len: int = 64,  # 1 hold == 2 tokens
+        smooth_labels: bool = False,
         shuffle_tokens: bool = True,
-        label_smoothing: bool = True,
-        prompt_size: float = 0.5,
+        prompt_size: float = 0.2,
+        subset: float = 1.0,
     ):
-        self.df = pd.read_csv(filename)
+        assert 0 < subset <= 1, f"Subset must be between 0 and 1, got {subset}"
+        self.df = pd.read_csv(filename).sample(frac=subset)
+        self.df["length"] = self.df["frames"].apply(lambda x: len(x) // 4 + 4)
+        self.bucket_shuffle()
         self.tokenizer = tokenizer
-        self.context_len = context_len
+        self.smooth_labels = smooth_labels
         self.shuffle_tokens = shuffle_tokens
-        self.label_smoothing = label_smoothing
         self.prompt_size = prompt_size
         self.eval = False
 
@@ -39,10 +43,8 @@ class KilterGPTDataset(Dataset):
             shuffle=self.shuffle_tokens,
         )
         n_tokens = tokenized.size(0)
-        prompt_size = int(n_tokens * self.prompt_size)
-        x = self.tokenizer.pad(tokenized[:prompt_size], self.context_len)
-        y = self.tokenizer.pad(tokenized, self.context_len)
-        return x, y
+        prompt_size = max(math.ceil(n_tokens * self.prompt_size), 5)
+        return tokenized[:prompt_size], tokenized
 
     def __getitem__(self, idx: int) -> tuple[torch.LongTensor, torch.Tensor]:
         if self.eval:
@@ -51,7 +53,7 @@ class KilterGPTDataset(Dataset):
             return self._get_item_train(idx)
 
     def _get_item_train(self, idx: int) -> tuple[torch.LongTensor, torch.Tensor]:
-        """Get a random contiguous sequence of tokens from the frames column. Pad left to context_len."""
+        """Get a training item. This will return a tuple of two tensors, x and y, where x is the input and y is the target."""
         row = self.df.iloc[idx]
         frames = row["frames"]
         tokenized = self.tokenizer.encode(
@@ -62,19 +64,33 @@ class KilterGPTDataset(Dataset):
         )
         x = tokenized[:-1]
         y = tokenized[1:]
-        if self.label_smoothing:
-            y = self._create_smoothed_labels(y)
-        x = self.tokenizer.pad(x, self.context_len)
-        y = self.tokenizer.pad(y, self.context_len)
+        if self.smooth_labels:
+            y = self.smooth_y(tokenized[1:])
         return x, y
 
-    def _create_smoothed_labels(self, y: torch.LongTensor) -> torch.FloatTensor:
-        """Make all holds equally valid."""
-        nopad = y[y != self.tokenizer.pad_token_id]
-        labels = torch.nn.functional.one_hot(nopad, num_classes=self.tokenizer.vocab_size).float()
-        # Positions of hold tokens in the y tensor
-        hold_indices = torch.tensor([x for x in range(2, nopad.size(0), 2)])
-        hold_values = nopad[hold_indices]
-        for idx, i in enumerate(hold_indices):
-            labels[i, hold_values[idx:]] = 1
-        return labels
+    def smooth_y(self, y: torch.Tensor) -> torch.Tensor:
+        smooth_y = F.one_hot(y, num_classes=self.tokenizer.vocab_size).to(torch.float32)
+        hold_positions = torch.arange(2, y.size(0) - 1, 2)
+        holds = y[hold_positions]
+        for i in range(len(holds)):
+            smooth_y[hold_positions[i], holds[i:]] = 1
+        return smooth_y
+
+    def shuffle(self):
+        """Just shuffle the dataframe"""
+        self.df = self.df.sample(frac=1).reset_index(drop=True)
+
+    def len_sort(self):
+        """Sort the dataframe by length of frames"""
+        self.df = self.df.sort_values(by="length", ascending=True).reset_index(drop=True)
+
+    def bucket_shuffle(self):
+        """Shuffle the bucket order and within the buckets.
+        A bucket is all sequences of the same length."""
+        self.shuffle()
+        self.df = pd.concat(
+            [group.sample(frac=1) for _, group in self.df.sample(frac=1).groupby("length", sort=False)]
+        )
+
+    def __repr__(self):
+        return f"KilterDataset of length {self.__len__()}"