+++ kiltergpt/models/gpt.py
@@ -1,12 +1,15 @@
 import math
+from pathlib import Path
 
 import lightning as L
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+import torchmetrics.functional as M
 from fastapi import FastAPI
 
 from ..utils import Plotter, WarmupCosineSchedule
+from ..utils.metrics import get_histogram, jaccard_similarity
 
 
 class LayerNorm(nn.Module):
@@ -47,7 +50,12 @@ class CausalSelfAttention(nn.Module):
 
         # causal self-attention; Self-attend: (B, nh, T, hs) x (B, nh, hs, T) -> (B, nh, T, T)
         y = F.scaled_dot_product_attention(
-            q, k, v, attn_mask=None, dropout_p=self.dropout if self.training else 0, is_causal=True
+            q,
+            k,
+            v,
+            attn_mask=None,
+            dropout_p=self.dropout if self.training else 0,
+            is_causal=True,
         )
         y = y.transpose(1, 2).contiguous().view(B, T, C)  # re-assemble all head outputs side by side
         y = self.resid_dropout(self.c_proj(y))
@@ -141,8 +149,8 @@ class GPTModel(L.LightningModule):
         self.config = config
         self.tokenizer = tokenizer
         self.model = GPT(self.config)
-        # self.model = torch.compile(self.model)
 
+    # TODO add tests
     def get_loss(self, logits, targets):
         B, C, V = logits.shape
         logits = logits.view(B * C, V)
@@ -158,7 +166,7 @@ class GPTModel(L.LightningModule):
         logits = self.model.forward(x)
         return logits
 
-    def shared_step(self, batch, name="train"):
+    def shared_step(self, batch, name: str):
         text, target = batch
         logits = self.forward(text)
         loss = self.get_loss(logits, target)
@@ -172,26 +180,82 @@ class GPTModel(L.LightningModule):
         return self.shared_step(batch, "val")
 
     def on_test_epoch_start(self) -> None:
-        self.test_generated = []
-        self.test_real = []
+        self.test_generated = {0.3: [], 0.5: [], 0.7: []}
+        self.test_real = {0.3: [], 0.5: [], 0.7: []}
         return super().on_test_epoch_start()
 
     def test_step(self, batch, batch_idx):
         prompts, targets = batch
-        for prompt, target in zip(prompts, targets, strict=True):
-            self.test_generated.append(self.generate(prompt, 0.2, 0.7).detach().cpu())
-            self.test_real.append(target.detach().cpu())
+        for temp in self.test_generated.keys():
+            for prompt, target in zip(prompts, targets, strict=True):
+                generated = self.generate(prompt, temp, p=0.8)
+                self.test_generated[temp].append(generated.detach().cpu())
+                self.test_real[temp].append(target.detach().cpu())
 
+    def _get_hist_pearson(self, generated, real):
+        """Calculate the spearman correlation between token distributions of real and genenerated sequences"""
+        hist_generated = get_histogram(generated, self.tokenizer.vocab_size)
+        hist_real = get_histogram(real, self.tokenizer.vocab_size)
+        hist_spearman = M.spearman_corrcoef(hist_generated, hist_real)
+        return hist_spearman
+
+    def _get_jaccard_similarity(self, generated, real):
+        generated_onehot = torch.stack([self.tokenizer.onehot(x) for x in generated])
+        real_onehot = torch.stack([self.tokenizer.onehot(x) for x in real])
+        jaccard = jaccard_similarity(generated_onehot, real_onehot)
+        return jaccard
+
+    def _get_num_possible_climbs(self, temp: float):
+        n_runs = 40
+        # test run, if the model is too small
+        if self.config.n_embed < 32:
+            n_runs = 4
+        climbs = set()
+        for _ in range(n_runs):
+            climb = self.generate_from_string("p1387r14", 40, "7a", temp, 0.7)
+            onehot = self.tokenizer.onehot(climb)
+            climbs.add(tuple(onehot.tolist()))
+        return len(climbs) / n_runs
+
+    def on_test_epoch_end(self) -> None:
+        for temp in self.test_generated.keys():
+            prefix = f"test/temp={temp}"
+            hist_spearman = self._get_hist_pearson(self.test_generated[temp], self.test_real[temp])
+            jaccard_similarity = self._get_jaccard_similarity(self.test_generated[temp], self.test_real[temp])
+            num_possible_climbs = self._get_num_possible_climbs(temp)
+            self.log_dict(
+                {
+                    f"{prefix}/hist_spearman": hist_spearman,
+                    f"{prefix}/jaccard_similarity": jaccard_similarity.mean(),
+                    f"{prefix}/num_possible_climbs": num_possible_climbs,
+                }
+            )
+
+    # TODO add tests
+    def on_train_epoch_end(self):
+        if self.config.only_train:
+            return super().on_train_epoch_end()
+        plotter = Plotter()
+        finish_hold = "p1387"
+        setup = [(30, "6a"), (40, "7a"), (50, "8a")]
+        for temp in [0.3, 0.5, 0.7]:
+            route_frames = [
+                self.generate_from_string(f"{finish_hold}r14", angle, grade, temperature=temp, p=0.8)
+                for angle, grade in setup
+            ]
+            route_images = [plotter.plot_climb(x, highlight=finish_hold) for x in route_frames]
+            captions = [f"{grade} @ {angle}, temp={temp}" for angle, grade in setup]
+            self.logger.log_image(key=f"test/temp={temp}/images", images=route_images, caption=captions)
+        return super().on_train_epoch_end()
+
+    # TODO add tests
     def configure_optimizers(self):
-        param_dict = {pn: p for pn, p in self.named_parameters()}
-        param_dict = {pn: p for pn, p in param_dict.items() if p.requires_grad}
-        decay_params = [p for n, p in param_dict.items() if p.dim() >= 2]
-        nodecay_params = [p for n, p in param_dict.items() if p.dim() < 2]
-        optim_groups = [
-            {"params": decay_params, "weight_decay": self.config.wd},
-            {"params": nodecay_params, "weight_decay": 0.0},
-        ]
-        optimizer = torch.optim.AdamW(optim_groups, lr=self.config.lr, betas=(0.9, 0.95), fused=True)
+        optimizer = torch.optim.AdamW(
+            self.parameters(),
+            lr=self.config.lr,
+            weight_decay=self.config.wd,
+            betas=(0.9, 0.95),
+        )
         scheduler = WarmupCosineSchedule(
             optimizer,
             self.config.total_steps // 10,
@@ -207,20 +271,18 @@ class GPTModel(L.LightningModule):
         return self.model.embed(x)
 
     def _sample_from_logits(self, logits: torch.Tensor, p: float = 1.0) -> torch.Tensor:
+        """Given logits of tokens, sample using top-p sampling"""
         if p < 1.0:
             # sort by probability, get cumulative probs
             sorted_logits, sorted_indices = torch.sort(logits, descending=True)
             cumulative_probs = torch.cumsum(F.softmax(sorted_logits, dim=-1), dim=-1)
-
             # Remove tokens with cumulative probability above the threshold
             sorted_indices_to_remove = cumulative_probs > p
             sorted_indices_to_remove[1:] = sorted_indices_to_remove[:-1].clone()
             sorted_indices_to_remove[0] = 0
-
             # Scatter sorted tensors to original indexing
             indices_to_remove = sorted_indices_to_remove.scatter(0, sorted_indices, sorted_indices_to_remove)
             logits[indices_to_remove] = float("-inf")
-
         # Sample from the filtered distribution
         probs = F.softmax(logits, dim=-1)
         next_prompt = torch.multinomial(probs, num_samples=1).to(self.device)
@@ -228,22 +290,34 @@ class GPTModel(L.LightningModule):
 
     def _generate_token(self, prompt: torch.Tensor, temperature: float = 0.2, p: float = 1.0) -> torch.Tensor:
         """Generate a single token
-        prompt: torch.LongTensor: A left-padded tensor of token ids
+        prompt: torch.LongTensor: input_ids
         """
         # TODO additionally check the number of start and finish tokens
-        penultimate_token = prompt[-1]
+        last_token = prompt[-1]
         logits = self.forward(prompt.unsqueeze(0)).squeeze(0)
         logits = logits[-1, :] / temperature  # Get logits for the last position
-        # After color token only hold or EOS token can be generated
-        if penultimate_token in self.tokenizer.color_token_ids.to(self.device):
+        # After color a token only hold or EOS token can be generated
+        if last_token in self.tokenizer.color_token_ids.to(self.device):
             mask = torch.zeros_like(logits, dtype=torch.bool)
+            previously_used = prompt[torch.isin(prompt, self.tokenizer.hold_token_ids.to(self.device))]
+            # Only allow hold tokens and EOS token
             mask[self.tokenizer.hold_token_ids] = True
             mask[self.tokenizer.eos_token_id] = True
+            # Previously used holds are not allowed (repetitions)
+            mask[previously_used] = False
             logits[~mask] = float("-inf")
-        # After hold token only color token can be generated
-        elif penultimate_token in self.tokenizer.hold_token_ids.to(self.device):
+        # After a hold token only color token can be generated
+        elif last_token in self.tokenizer.hold_token_ids.to(self.device):
             mask = torch.zeros_like(logits, dtype=torch.bool)
             mask[self.tokenizer.color_token_ids] = True
+            # not more than 2 starts
+            # TODO check that this works
+            if (prompt == self.tokenizer.start_token_id).to(torch.long).sum() >= 2:
+                mask[self.tokenizer.start_token_id] = False
+            # not more than 2 finishes
+            if (prompt == self.tokenizer.finish_token_id).to(torch.long).sum() >= 2:
+                mask[self.tokenizer.finish_token_id] = False
+            # All not allowed tokens set prob to 0
             logits[~mask] = float("-inf")
         return self._sample_from_logits(logits, p)
 
@@ -254,10 +328,11 @@ class GPTModel(L.LightningModule):
             context = prompt[-self.config.context_len :]
             next_prompt = self._generate_token(context, temperature, p)
             prompt = torch.cat((prompt, next_prompt), dim=0)
-            # Stop when you get to 2x length of context length
-            if len(prompt) > self.config.context_len * 2 - 1:
+            # Stop when you get to full context window (30 holds)
+            if prompt.size(0) >= self.config.context_len - 1:
                 prompt = torch.cat(
-                    (prompt, torch.tensor(self.tokenizer.eos_token_id, device=self.device).unsqueeze(0)), dim=0
+                    (prompt, torch.tensor(self.tokenizer.eos_token_id, device=self.device).unsqueeze(0)),
+                    dim=0,
                 )
                 break
         return prompt
@@ -268,27 +343,23 @@ class GPTModel(L.LightningModule):
         frames: str,
         angle: int,
         grade: str,
-        temperature: float = 0.2,
-        p: float = 0.7,
+        temperature: float = 0.7,
+        p: float = 0.8,
     ) -> str:
         """Generate a climb from a string of frames, angle, and grade"""
-        tokenized = self.tokenizer.encode(frames, angle, grade, pad=self.config.context_len, eos=False).to(self.device)
+        tokenized = self.tokenizer.encode(frames, angle, grade, eos=False).to(self.device)
         generated = self.generate(tokenized, temperature, p)
         return self.tokenizer.decode(generated, clean=True)[0]
 
     @staticmethod
-    def load_from_wandb(wandb_model_name: str) -> "GPTModel":
+    def load_from_wandb(model_name: str, repo_name: str = "ilsenatorov/model-registry") -> "GPTModel":
         """Use self.load_from_checkpoint to download model weights from wandb. Looks for models in ilsenatorov/kilter-gpt"""
-        import os
+        import wandb
 
-        file_path = f"artifacts/{wandb_model_name}/model.ckpt"
-        if not os.path.exists(file_path):
-            import wandb
-
-            api = wandb.Api()
-            artifact = api.artifact(f"ilsenatorov/kilter-gpt/{wandb_model_name}")
-            artifact.download()
-        return GPTModel.load_from_checkpoint(file_path)
+        api = wandb.Api()
+        artifact = api.artifact(f"{repo_name}/{model_name}")
+        artifact_dir = artifact.download()
+        return GPTModel.load_from_checkpoint(f"{artifact_dir}/model.ckpt")
 
     def get_fastapi_app(self) -> FastAPI:
         """Return a FastAPI app that serves the model. Can be launched with gunicorn."""