+++ tests/test_model.py
@@ -16,9 +16,9 @@ def sample_config():
         epochs=3,
         lr=6e-4,
         wd=1e-1,
-        n_head=4,
-        n_layer=4,
-        n_embed=128,
+        n_head=2,
+        n_layer=2,
+        n_embed=4,
         context_len=64,
         dropout=0.2,
         bias=False,
@@ -34,12 +34,11 @@ def test_forward_pass(sample_config):
         "p1234r12p1345r13p1423r14p1243r15",
         40,
         "7a",
-        pad=sample_config.context_len,
         shuffle=True,
     )
     sample_batch = sample_input.unsqueeze(0).repeat(2, 1)
     output = model(sample_batch)
-    assert output.shape == (2, 64, sample_config.vocab_size)
+    assert output.shape == (2, sample_input.size(0), sample_config.vocab_size)
 
 
 def test_generate(sample_config):
@@ -47,7 +46,25 @@ def test_generate(sample_config):
     model = GPTModel(sample_config, tokenizer)
     sample_prompt = "p1234r12"
     generated = model.generate_from_string(sample_prompt, 40, "7a")
-    # FIXME add sensible asserts
+    assert sample_prompt in generated
+
+
+def test_test_step(sample_config):
+    tokenizer = Tokenizer()
+    model = GPTModel(sample_config, tokenizer)
+    model.on_test_epoch_start()
+    sample_input = tokenizer.encode(
+        "p1234r12p1345r13p1423r14p1243r15",
+        40,
+        "7a",
+        shuffle=True,
+        eos=False,
+    )
+    sample_batch = sample_input.unsqueeze(0).repeat(2, 1)
+    model.test_step((sample_batch, sample_batch), 0)
+    assert len(model.test_generated[0.3]) == 2
+    assert len(model.test_real[0.3]) == 2
+    model.on_test_epoch_end()
 
 
 def test_app(sample_config):